// src/openapi/models/PagingSimplifiedAlbumObject.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* eslint-disable */

import type { PagingObject } from './PagingObject';
import type { SimplifiedAlbumObject } from './SimplifiedAlbumObject';

export type PagingSimplifiedAlbumObject = (PagingObject & {
items?: Array<SimplifiedAlbumObject>;
});